refactor cli into lib function
