+++ src/lib.rs
@@ -11,7 +11,7 @@ use serde_json::Value;
 use std::cell::RefCell;
 use std::fs::{self, File};
 use std::io::{self, BufRead};
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 use std::time::Instant;
 use tinytemplate::TinyTemplate;
 
@@ -27,7 +27,7 @@ mod types;
 pub use types::{
     ArtifactFlags, CollectiveSchedule, CollectivesParityReport, Diagnostics, DivergenceFlags,
     DivergenceGroup, ExecOrderSummary, GraphAnalysis, GraphCollectivesParity, GraphRuntime,
-    RankMetaData, RuntimeAnalysis, RuntimeRankDetail,
+    MultiRankContext, RankMetaData, RuntimeAnalysis, RuntimeRankDetail,
 };
 
 pub use execution_order::{
@@ -1288,6 +1288,392 @@ pub fn generate_multi_rank_html(
     let landing_page_path = out_path.join("index.html");
 
     Ok((landing_page_path, html))
+}
+
+/// Generate multi-rank landing and aggregated artifacts from pre-parsed rank outputs.
+///
+/// This mirrors the diagnostics and file emission performed by the CLI `--all-ranks-html` path.
+/// It expects that per-rank outputs already exist under `out_dir/rank_<n>` and will:
+/// - Combine Chromium events into `chromium_events.json` (if any)
+/// - Write `runtime_estimations.json` and `chromium_trace_with_runtime.json` (if any)
+/// - Write `collective_schedules.json` (if any) and per-rank `collectives_parity.json`
+/// - Analyze runtime deltas and exec order summary
+/// - Render `index.html` landing page using the same template and context
+pub fn generate_multi_rank_landing(
+    cfg: &ParseConfig,
+    ctx: &MultiRankContext,
+    out_dir: &Path,
+) -> anyhow::Result<()> {
+    use std::fs;
+
+    // Parse ranks from context and ensure sorted numerically like the CLI
+    let mut rank_nums: Vec<u32> = ctx
+        .ranks
+        .iter()
+        .filter_map(|s| s.parse::<u32>().ok())
+        .collect();
+    rank_nums.sort_unstable();
+    let sorted_ranks: Vec<String> = rank_nums.iter().map(|r| r.to_string()).collect();
+
+    let out_path: PathBuf = out_dir.to_path_buf();
+
+    // Collect compile ids and cache sequences per rank, and chromium events at root
+    let mut rank_metadata: FxHashMap<u32, RankMetaData> = FxHashMap::default();
+    let mut all_chromium_events: Vec<serde_json::Value> = Vec::new();
+
+    for &rank_num in &rank_nums {
+        let subdir = out_path.join(format!("rank_{rank_num}"));
+        let chromium_events_path = subdir.join("chromium_events.json");
+        let compile_dir_json = subdir.join("compile_directory.json");
+
+        // extract compile IDs and cache sequence from compile_directory.json
+        let mut compile_ids: FxHashSet<String> = FxHashSet::default();
+        let content = fs::read_to_string(&compile_dir_json)?;
+        let mut artifact_entries: Vec<(u64, String)> = Vec::new();
+
+        if let Ok(serde_json::Value::Object(map)) =
+            serde_json::from_str::<serde_json::Value>(&content)
+        {
+            for (key, val) in map.iter() {
+                if key != "unknown" && !key.starts_with("unknown_") {
+                    compile_ids.insert(key.clone());
+                }
+                if let Some(arr) = val.get("artifacts").and_then(|v| v.as_array()) {
+                    for art in arr {
+                        let suffix = art.get("suffix").and_then(|s| s.as_str()).unwrap_or("");
+                        if suffix.is_empty() {
+                            continue;
+                        }
+                        if let Some(num) = art.get("number").and_then(|n| n.as_u64()) {
+                            artifact_entries.push((num, suffix.to_string()));
+                        }
+                    }
+                }
+            }
+        }
+
+        artifact_entries.sort_by_key(|(n, _)| *n);
+        let cache_sequence: String = artifact_entries.into_iter().map(|(_, s)| s).collect();
+
+        rank_metadata.insert(
+            rank_num,
+            RankMetaData {
+                rank: rank_num,
+                compile_ids,
+                cache_sequence,
+            },
+        );
+
+        // collect chromium events for each rank
+        if chromium_events_path.exists() {
+            let events = read_chromium_events_with_pid(&chromium_events_path, rank_num)?;
+            all_chromium_events.extend(events);
+        }
+    }
+
+    // Determine if there is any divergence in compile IDs across ranks
+    let compile_id_divergence =
+        if let Some(first) = rank_nums.iter().filter_map(|r| rank_metadata.get(r)).next() {
+            rank_metadata
+                .values()
+                .any(|md| md.compile_ids != first.compile_ids)
+        } else {
+            false
+        };
+
+    // Group ranks by their cache hit/miss sequence
+    let cache_seq_groups: FxHashMap<String, Vec<u32>> =
+        rank_metadata
+            .into_values()
+            .fold(FxHashMap::default(), |mut acc, md| {
+                acc.entry(md.cache_sequence).or_default().push(md.rank);
+                acc
+            });
+
+    // Build groups describing cache hit/miss patterns per rank
+    let cache_divergence_groups: Vec<DivergenceGroup> = if cache_seq_groups.len() > 1 {
+        cache_seq_groups
+            .iter()
+            .map(|(seq, ranks_vec)| {
+                let mut sorted = ranks_vec.clone();
+                sorted.sort_unstable();
+                DivergenceGroup {
+                    sequence: seq.clone(),
+                    ranks: sorted
+                        .iter()
+                        .map(|r| r.to_string())
+                        .collect::<Vec<_>>()
+                        .join(", "),
+                }
+            })
+            .collect()
+    } else {
+        Vec::new()
+    };
+
+    // combine chromium events from all ranks
+    if !all_chromium_events.is_empty() {
+        let combined_chromium_path = out_path.join("chromium_events.json");
+        let combined_events_json = serde_json::to_string_pretty(&all_chromium_events)?;
+        fs::write(combined_chromium_path, combined_events_json)?;
+    }
+
+    // Process runtime estimations from all ranks
+    let runtime_estimations = crate::parsers::read_runtime_estimations(&out_path, &rank_nums)?;
+    if !runtime_estimations.is_empty() {
+        let runtime_path = out_path.join("runtime_estimations.json");
+        fs::write(
+            &runtime_path,
+            serde_json::to_string_pretty(&runtime_estimations)?,
+        )?;
+        println!("Runtime estimations: {}", runtime_path.display());
+
+        // Generate runtime trace events in a single pass
+        let mut runtime_events: Vec<serde_json::Value> = Vec::new();
+        let mut pid_set: FxHashSet<u32> = FxHashSet::default();
+        let mut thread_names: FxHashMap<(u32, u32), String> = FxHashMap::default();
+
+        // Concise, deterministic 32-bit TID from (rank, graph)
+        let calc_tid = |rank: u32, graph: &str| -> u32 {
+            use std::hash::{Hash, Hasher};
+            let mut h = fxhash::FxHasher::default();
+            (rank, graph).hash(&mut h);
+            (h.finish() & 0xFFFF_FFFF) as u32
+        };
+
+        for gr in &runtime_estimations {
+            pid_set.insert(gr.rank);
+            let tid = calc_tid(gr.rank, &gr.graph);
+            thread_names
+                .entry((gr.rank, tid))
+                .or_insert_with(|| gr.graph.clone());
+
+            let mut time_offset_us: u64 = 0;
+            for op in &gr.ops {
+                let dur_us = (op.estimated_runtime_ns / 1000.0).ceil().max(1.0) as u64;
+                runtime_events.push(serde_json::json!({
+                    "name": op.name,
+                    "ph": "X",
+                    "ts": time_offset_us,
+                    "dur": dur_us,
+                    "pid": gr.rank,
+                    "tid": tid,
+                    "cat": "runtime",
+                    "args": {
+                        "graph": gr.graph,
+                        "rank": gr.rank,
+                        "runtime_ns": op.estimated_runtime_ns as u64
+                    }
+                }));
+                time_offset_us += dur_us;
+            }
+        }
+
+        let mut all_events: Vec<serde_json::Value> = runtime_events;
+
+        // Emit process (rank) metadata in ascending pid order
+        let mut pids: Vec<u32> = pid_set.into_iter().collect();
+        pids.sort_unstable();
+        for pid in pids.into_iter() {
+            all_events.extend([
+                serde_json::json!({
+                    "name": "process_name",
+                    "ph": "M",
+                    "pid": pid,
+                    "args": {"name": format!("Rank {}", pid)}
+                }),
+                serde_json::json!({
+                    "name": "process_sort_index",
+                    "ph": "M",
+                    "pid": pid,
+                    "args": {"sort_index": pid as i64}
+                }),
+            ]);
+        }
+
+        // Emit thread names sorted by graph name within each pid
+        let mut threads_by_pid: FxHashMap<u32, Vec<(u32, String)>> = FxHashMap::default();
+        for ((pid, tid), graph_name) in thread_names.into_iter() {
+            threads_by_pid
+                .entry(pid)
+                .or_default()
+                .push((tid, graph_name));
+        }
+        let mut pids_for_threads: Vec<u32> = threads_by_pid.keys().copied().collect();
+        pids_for_threads.sort_unstable();
+        for pid in pids_for_threads {
+            let entries = threads_by_pid.remove(&pid).unwrap_or_default();
+            for (idx, (tid, graph_name)) in entries.into_iter().enumerate() {
+                all_events.extend([
+                    serde_json::json!({
+                        "name": "thread_name",
+                        "ph": "M",
+                        "pid": pid,
+                        "tid": tid,
+                        "args": {"name": format!("graph {}", graph_name)}
+                    }),
+                    serde_json::json!({
+                        "name": "thread_sort_index",
+                        "ph": "M",
+                        "pid": pid,
+                        "tid": tid,
+                        "args": {"sort_index": idx as i64}
+                    }),
+                ]);
+            }
+        }
+
+        fs::write(
+            out_path.join("chromium_trace_with_runtime.json"),
+            serde_json::to_string_pretty(&all_events)?,
+        )?;
+    }
+
+    // Analyze graph runtime deltas across ranks
+    let runtime_analysis = if !runtime_estimations.is_empty() {
+        analyze_graph_runtime_deltas(&runtime_estimations)
+    } else {
+        None
+    };
+
+    // Process collective schedules from all ranks
+    let collective_schedules = crate::parsers::read_collective_schedules(&out_path, &rank_nums)?;
+    if !collective_schedules.is_empty() {
+        let schedules_path = out_path.join("collective_schedules.json");
+        fs::write(
+            &schedules_path,
+            serde_json::to_string_pretty(&collective_schedules)?,
+        )?;
+        println!("Collective schedules: {}", schedules_path.display());
+    }
+
+    crate::parsers::check_collectives_parity(&out_path, &rank_nums)?;
+
+    // Process tensor meta fingerprints from all ranks
+    let tensor_meta = crate::parsers::read_tensor_meta_fingerprints(&out_path, &rank_nums)?;
+    let mut tensor_meta_groups: FxHashMap<String, Vec<u32>> = FxHashMap::default();
+    if !tensor_meta.is_empty() {
+        use std::collections::HashMap;
+        // rank -> sorted list of (graph_id, fingerprint)
+        let mut by_rank: HashMap<u32, Vec<(String, String)>> = HashMap::new();
+        for tm in &tensor_meta {
+            by_rank
+                .entry(tm.rank)
+                .or_default()
+                .push((tm.graph.clone(), tm.fingerprint.clone()));
+        }
+        for (&rank, entries) in &mut by_rank {
+            // sort by graph id to make cross-rank concatenation consistent
+            let mut entries_sorted = entries.clone();
+            entries_sorted.sort_by(|a, b| a.0.cmp(&b.0));
+            let signature = entries_sorted
+                .into_iter()
+                .map(|(_, fp)| fp)
+                .collect::<Vec<_>>()
+                .join(",");
+            tensor_meta_groups.entry(signature).or_default().push(rank);
+        }
+    }
+
+    let tensor_meta_divergence_groups: Vec<DivergenceGroup> = if tensor_meta_groups.len() > 1 {
+        tensor_meta_groups
+            .iter()
+            .map(|(seq, ranks_vec)| {
+                let mut sorted = ranks_vec.clone();
+                sorted.sort_unstable();
+                DivergenceGroup {
+                    sequence: seq.clone(),
+                    ranks: sorted
+                        .iter()
+                        .map(|r| r.to_string())
+                        .collect::<Vec<_>>()
+                        .join(", "),
+                }
+            })
+            .collect()
+    } else {
+        Vec::new()
+    };
+
+    // Group ranks by their collective op sequence
+    let mut collective_seq_groups: FxHashMap<String, Vec<u32>> = FxHashMap::default();
+    if !collective_schedules.is_empty() {
+        for &rank in &rank_nums {
+            let ops_concat: String = collective_schedules
+                .iter()
+                .filter(|s| s.rank == rank)
+                .flat_map(|s| s.ops.clone())
+                .collect::<Vec<_>>()
+                .join(",");
+            collective_seq_groups
+                .entry(ops_concat)
+                .or_default()
+                .push(rank);
+        }
+    }
+
+    let collective_divergence_groups: Vec<DivergenceGroup> = if collective_seq_groups.len() > 1 {
+        collective_seq_groups
+            .iter()
+            .map(|(seq, ranks_vec)| {
+                let mut sorted = ranks_vec.clone();
+                sorted.sort_unstable();
+                DivergenceGroup {
+                    sequence: seq.clone(),
+                    ranks: sorted
+                        .iter()
+                        .map(|r| r.to_string())
+                        .collect::<Vec<_>>()
+                        .join(", "),
+                }
+            })
+            .collect()
+    } else {
+        Vec::new()
+    };
+
+    println!(
+        "Multi-rank report generated under {}\nIndividual pages: rank_*/index.html",
+        out_path.display()
+    );
+
+    let exec_order_summary = build_exec_order_summary(&out_path, &rank_nums, &collective_schedules);
+
+    let diagnostics = Diagnostics {
+        divergence: DivergenceFlags {
+            cache: cache_seq_groups.len() > 1,
+            collective: collective_seq_groups.len() > 1,
+            tensor_meta: tensor_meta_groups.len() > 1,
+        },
+        artifacts: ArtifactFlags {
+            runtime_trace: !runtime_estimations.is_empty(),
+        },
+        analysis: runtime_analysis,
+        cache_groups: cache_divergence_groups.clone(),
+        collective_groups: collective_divergence_groups.clone(),
+        tensor_meta_groups: tensor_meta_divergence_groups.clone(),
+        exec_order: exec_order_summary,
+    };
+
+    // Emit landing page identical to CLI
+    let has_chromium_events = !all_chromium_events.is_empty();
+    let show_desync_warning = compile_id_divergence
+        || diagnostics.divergence.cache
+        || diagnostics.divergence.collective
+        || diagnostics.divergence.tensor_meta;
+
+    let (landing_page_path, landing_html) = generate_multi_rank_html(
+        &out_path,
+        sorted_ranks,
+        cfg,
+        has_chromium_events,
+        show_desync_warning,
+        compile_id_divergence,
+        diagnostics,
+    )?;
+    fs::write(&landing_page_path, landing_html)?;
+
+    Ok(())
 }
 
 /// Build ExecOrderSummary from artifacts under out_path for the given ranks
